NamedVaue util class renamed to NamedString (minus generics)
