+++ java-client/src/main/java/co/elastic/clients/util/NamedString.java
@@ -1,0 +1,66 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package co.elastic.clients.util;
+
+import java.util.Objects;
+
+public class NamedString {
+
+    private final String name;
+
+    private final String value;
+
+    public NamedString(String name, String value) {
+        this.name = name;
+        this.value = value;
+    }
+
+    public String name() {
+        return this.name;
+    }
+
+    public String value() {
+        return this.value;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (this == other) return true;
+        if (!(other instanceof NamedString)) return false;
+        NamedString namedString = (NamedString) other;
+        return Objects.equals(name(), namedString.name()) && Objects.equals(value(), namedString.value());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name(), value());
+    }
+
+    @Override
+    public String toString() {
+        if (name == null || value == null) {
+            return "";
+        }
+        else {
+            return String.format("%s: %s", name(), value());
+        }
+    }
+
+}
